Pass req and attribute to required rule
If replacing the default required rule by registering a custom rule of the same name, the replacement may need additional context other than just the value to determine if a field is required, but currently only the value is passed to the required rule.  Just changing the calls to the required rule in the various other required rule variants to pass req and attribute.